requirement.txt
google.generativeai
langchain
python-dotenv